requirements.txt
# Python Dependencies for YouTube Blog Automation System

# Core CrewAI Framework
crewai
crewai-tools

# Google Gemini LLM
google-generativeai
litellm

# YouTube Integration
youtube-transcript-api
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# Stock Image APIs
requests
urllib3

# Environment and Configuration
python-dotenv
pyyaml

# Content Processing
markdown
beautifulsoup4

# Image Processing (optional)
Pillow

# Utilities
pathlib2
rich # For better console output
click  # For CLI improvements

# Logging and Monitoring
loguru

# Date/Time Utilities
python-dateutil

# HTTP Clients
httpx

# Data Validation
pydantic
